create list_item.rb for Single-responsiblity Principle
